framework for clientmanager?
