+++ src/test/java/com/snowflake/kafka/connector/internal/IngestSdkProviderTest.java
@@ -20,6 +20,8 @@ package com.snowflake.kafka.connector.internal;
 import com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig;
 import com.snowflake.kafka.connector.Utils;
 import java.util.Map;
+
+import com.snowflake.kafka.connector.internal.ingestsdk.IngestSdkProvider;
 import net.snowflake.ingest.streaming.SnowflakeStreamingIngestClient;
 import net.snowflake.ingest.utils.SFException;
 import org.apache.kafka.connect.errors.ConnectException;
